+++ tests/__main__.py
@@ -5,9 +5,13 @@
 import unittest
 
 
-if __name__ == '__main__':
+def main():
     loader = unittest.TestLoader()
     start_dir = 'tests'
     suite = loader.discover(start_dir)
     runner = unittest.TextTestRunner()
     runner.run(suite)
+
+
+if __name__ == '__main__':
+    main()